+++ CalculationEngine/HouseholdElements/CalcAffordanceBase.cs
@@ -74,7 +74,7 @@ namespace CalculationEngine.HouseholdElements
         protected readonly int SubAffordanceStartFrame;
 
         protected CalcAffordanceBase([JetBrains.Annotations.NotNull] string pName, [JetBrains.Annotations.NotNull] CalcLocation loc, [JetBrains.Annotations.NotNull][ItemNotNull] List<CalcDesire> satisfactionvalues, int miniumAge, int maximumAge,
-            PermittedGender permittedGender, bool needsLight, bool randomEffect, [JetBrains.Annotations.NotNull] string pAffCategory, bool isInterruptable, bool isInterrupting, ActionAfterInterruption actionAfterInterruption, int weight,
+            PermittedGender permittedGender, bool needsLight, bool randomEffect, [JetBrains.Annotations.NotNull] string pAffCategory, bool isInterruptable, bool isInterrupting, ActionAfterInterruption actionAfterInterruption, double weight,
             bool requireAllAffordances, CalcAffordanceType calcAffordanceType, StrGuid guid, BodilyActivityLevel bodilyActivityLevel, [JetBrains.Annotations.NotNull] CalcRepo calcRepo,
             HouseholdKey householdKey, List<DeviceEnergyProfileTuple> energyProfiles, ColorRGB affordanceColor, string sourceTrait, string? timeLimitName, [JetBrains.Annotations.NotNull] CalcVariableRepository variableRepository,
             [JetBrains.Annotations.NotNull][ItemNotNull] List<CalcAffordanceVariableOp> variableOps, [JetBrains.Annotations.NotNull][ItemNotNull] List<VariableRequirement> variableRequirements) : base(pName, guid)
@@ -185,7 +185,7 @@ namespace CalculationEngine.HouseholdElements
 
         public string? TimeLimitName { get; }
 
-        public int Weight { get; }
+        public double Weight { get; }
         public HouseholdKey HouseholdKey { get; }
 
         public override string ToString() => "Affordance:" + Name;
